# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UnbindMacRequest": {
        "properties": {
          "mac_address": {
            "title": "Mac Address",
            "type": "string"
          }
        },
        "required": [
          "mac_address"
        ],
        "title": "UnbindMacRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WifiLoginChallengeRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "student_id": {
            "anyOf": [
              {
                "maxLength": 13,
                "minLength": 8,
                "pattern": "^[Uu]?\\d{8,12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          }
        },
        "title": "WifiLoginChallengeRequest",
        "type": "object"
      },
      "WifiLoginRequest": {
        "properties": {
          "captcha_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captcha Code"
          },
          "challenge_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Token"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "WifiLoginRequest",
        "type": "object"
      },
      "WifiStandaloneLoginRequest": {
        "properties": {
          "captcha_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captcha Code"
          },
          "challenge_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Token"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "student_id": {
            "maxLength": 13,
            "minLength": 8,
            "pattern": "^[Uu]?\\d{8,12}$",
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "password"
        ],
        "title": "WifiStandaloneLoginRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "USTB Manager",
      "url": "https://github.com/yourusername/ustb-manager"
    },
    "description": "\n## USTB Manager - 北科大教务管理系统API\n\n这是一个用于访问北京科技大学教务系统的API服务，提供以下功能：\n\n### 🔐 认证模块 (Authentication)\n- **二维码登录**: 使用微信扫码快速登录\n- **短信验证码登录**: 通过手机号接收验证码登录\n- **Cookie登录**: 使用浏览器Cookie直接登录（高级用户）\n- **会话管理**: 自动管理用户会话，支持后端重启后恢复\n\n### 📊 成绩查询模块 (Grades)\n- **成绩列表**: 支持分页、筛选的成绩查询\n- **GPA计算**: 自动计算总GPA和加权GPA\n- **学生信息**: 获取学生基本信息和完整档案\n- **培养方案**: 查看专业培养方案和课程要求\n- **学业进度**: 查询必修课完成情况和学分统计\n\n### 🔒 安全特性\n- HttpOnly Cookie保护\n- Session自动过期和清理\n- Cookie有效性验证\n- 支持CORS跨域访问\n\n### 📝 使用说明\n1. 首先调用认证接口（QR/SMS/Cookie）完成登录\n2. 登录成功后会自动设置session cookie\n3. 后续请求会自动携带cookie进行认证\n4. 可以调用 `/api/auth/status` 检查登录状态\n5. 使用 `/api/auth/logout` 退出登录\n\n### 🌐 环境信息\n- 开发环境: http://localhost:8000\n- API前缀: `/api`\n- 文档地址: `/docs` (Swagger UI) 或 `/redoc` (ReDoc)\n\n### 📞 技术支持\n如有问题，请查看各接口的详细文档说明。\n    ",
    "license": {
      "name": "MIT License"
    },
    "title": "USTB Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/wifi/bills": {
      "get": {
        "description": "获取历史账单\n\nArgs:\n    year: 年份（可选，默认当前年份，范围2000-2100）\n\nReturns:\n    {\n        \"year\": 查询年份,\n        \"summary\": {\n            \"total_time\": 总使用时长(分钟),\n            \"total_flow\": 总使用流量(MB),\n            \"base_fee\": 基本月租(元),\n            \"usage_fee\": 时长/流量计费(元)\n        },\n        \"bills\": [{\n            \"start_time\": 账单开始时间,\n            \"end_time\": 账单结束时间,\n            \"package\": 套餐,\n            \"base_fee\": 基本月租,\n            \"usage_fee\": 时长/流量计费,\n            \"duration_minutes\": 使用时长,\n            \"used_flow_mb\": 使用流量,\n            \"bill_time\": 出账时间\n        }]\n    }",
        "operationId": "get_bills_api_wifi_bills_get",
        "parameters": [
          {
            "description": "年份",
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "description": "年份",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bills",
        "tags": [
          "校园网管理"
        ]
      }
    },
...
    "/api/wifi/login": {
      "post": {
        "description": "登录校园网（使用学号 + 校园网密码）\n\n学号从教务系统 session 获取，只需要提供密码",
        "operationId": "wifi_login_api_wifi_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WifiLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wifi Login",
        "tags": [
          "校园网管理"
        ]
      }
    },
    "/api/wifi/login/challenge": {
      "post": {
        "description": "获取校园网验证码挑战。",
        "operationId": "wifi_login_challenge_api_wifi_login_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/WifiLoginChallengeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wifi Login Challenge",
        "tags": [
          "校园网管理"
        ]
      }
    },
...
    "/api/wifi/mac-vendor": {
      "get": {
        "description": "查询MAC地址厂商（离线数据库）\n\n使用IEEE OUI数据库（约38000条记录）查询MAC地址对应的设备厂商。\n自动识别随机MAC地址（iOS/Android隐私保护功能生成的本地管理地址）。\n\nArgs:\n    mac: MAC地址，支持多种格式（XX-XX-XX-XX-XX-XX 或 XX:XX:XX:XX:XX:XX 或 XXXXXXXXXXXX）\n\nReturns:\n    {\n        \"vendor\": 厂商名称（随机MAC返回\"随机MAC\"）,\n        \"is_random\": 是否为随机MAC地址\n    }",
        "operationId": "get_mac_vendor_api_wifi_mac_vendor_get",
        "parameters": [
          {
            "description": "MAC地址",
            "in": "query",
            "name": "mac",
            "required": true,
            "schema": {
              "description": "MAC地址",
              "title": "Mac",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mac Vendor",
        "tags": [
          "校园网管理"
        ]
      }
    },
    "/api/wifi/payments": {
      "get": {
        "description": "获取充值明细\n\nArgs:\n    start_date: 开始日期（可选，默认一年前，格式 YYYY-MM-DD）\n    end_date: 结束日期（可选，默认今天，格式 YYYY-MM-DD）\n\nReturns:\n    {\n        \"start_date\": 开始日期,\n        \"end_date\": 结束日期,\n        \"total_amount\": 充值总金额(元),\n        \"payments\": [{\n            \"pay_time\": 交费时间,\n            \"amount\": 金额,\n            \"pay_type\": 交费类型,\n            \"terminal\": 受理终端,\n            \"remark\": 备注\n        }]\n    }",
        "operationId": "get_payment_records_api_wifi_payments_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "title": "Start Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "title": "End Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Payment Records",
        "tags": [
          "校园网管理"
        ]
      }
    },
    "/api/wifi/standalone-login": {
      "post": {
        "description": "独立登录校园网（不需要教务系统登录）\n\n需要提供学号和密码，登录成功后设置 wifi_student_id cookie",
        "operationId": "wifi_standalone_login_api_wifi_standalone_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WifiStandaloneLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Wifi Standalone Login",
        "tags": [
          "校园网管理"
        ]
      }
    },
...
    "/api/wifi/unbind-mac": {
      "post": {
        "description": "解绑指定的MAC地址\n\nArgs:\n    mac_address: 要解绑的MAC地址（格式：XX-XX-XX-XX-XX-XX 或 XXXXXXXXXXXX）\n\nReturns:\n    {\n        \"success\": 是否成功,\n        \"message\": 消息\n    }",
        "operationId": "unbind_mac_address_api_wifi_unbind_mac_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindMacRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unbind Mac Address",
        "tags": [
          "校园网管理"
        ]
      }
    }
  },
...
}